requirements.txt
python-dotenv
requests
pandas
alive-progress
click